requirements.txt
streamlit==1.31.0
streamlit-extras==0.3.6
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.18.0
pydantic==2.5.2
bcrypt==4.0.1
uuid==1.30
# データベース関連
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
# 認証関連
PyJWT==2.8.0
# sqlite3はPython標準ライブラリのため、追加不要